# repositories/private-registry-api/api/openapi.yaml
openapi: 3.1.0
info:
  title: Registry API
  version: 1.0.0
  description: |
    Authenticated catalog, governance, event-stream, and operational interfaces for Registry.
    Package binaries remain in Artifactory and are never served by this API. Catalog responses are
    filtered by the caller's Microsoft Entra APM entitlements; unauthorized and nonexistent package
    identities intentionally return the same response.
servers:
  - url: /
    description: Same-origin Registry deployment
security:
  - oidcSession: []
...
paths:
...
  /api/v1/catalog/packages:
    get:
      tags: [Catalog]
      operationId: listCatalogPackages
      summary: Search and browse the authorized catalog
      parameters:
        - name: q
          in: query
          description: Case-insensitive search across namespace, name, title, and description.
          schema: {type: string}
        - name: kind
          in: query
          schema: {type: string, enum: [module, provider]}
        - name: provider
          in: query
          description: Module target or provider name.
          schema: {type: string}
        - $ref: '#/components/parameters/ApmId'
        - name: lifecycle
          in: query
          schema:
            type: string
            enum: [draft, candidate, approved, maintenance, deprecated, revoked, archived]
        - name: approval
          in: query
          schema: {type: string, enum: [approved, rejected, waived]}
        - name: risk
          in: query
          schema: {type: string, enum: [low, medium, high, critical]}
        - name: sort
          in: query
          description: Relevance applies when q is present; without q it is normalized to updated.
          schema:
            type: string
            enum: [relevance, updated, name, risk]
            default: updated
        - name: cursor
          in: query
          description: Opaque cursor returned by the previous page; it must be used with the same sort.
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 25}
      responses:
        '200':
          description: Authorized catalog page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/IdentityUnavailable'
  /api/v1/catalog/counts:
    get:
      tags: [Catalog]
      operationId: getCatalogCounts
      summary: Return truthful authorized package counts
      responses:
        '200':
          description: Counts include only packages visible to the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogCounts'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/IdentityUnavailable'
...
  /api/v1/catalog/events:
    get:
      tags: [Event stream]
      operationId: streamCatalogEvents
      summary: Stream authorized catalog changes
      description: Events never include APM assignment lists. Clients should invalidate affected catalog queries.
      parameters:
        - $ref: '#/components/parameters/ApmId'
      responses:
        '200':
          description: Server-sent event stream containing connected and catalog-change events.
          content:
            text/event-stream:
              schema: {type: string}
              examples:
                catalogChange:
                  value: |-
                    event: catalog-change
                    id: provider/platform/cloud:2026-07-22T04:00:00Z
                    data: {"package_id":"provider/platform/cloud","change_type":"indexed","occurred_at":"2026-07-22T04:00:00Z"}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/IdentityUnavailable'
...
  /internal/webhooks/jfrog:
    post:
      tags: [Webhook]
      operationId: receiveJfrogWebhook
      summary: Accept a signed JFrog artifact or property event
      description: |
        This route bypasses interactive OIDC authentication but validates the raw-body HMAC signature,
        configured JPD origin, subscription identifier, repository allowlist, content type, and payload size.
        The signing secret and delegated credentials are never returned by any API.
      security: []
      parameters:
        - name: X-JFrog-Signature
          in: header
          required: true
          description: HMAC-SHA256 over the exact request bytes, encoded as hex, sha256-prefixed hex, or base64.
          schema: {type: string, writeOnly: true}
        - name: X-JFrog-Origin
          in: header
          required: true
          schema: {type: string}
        - name: X-JFrog-Subscription-Id
          in: header
          required: true
          schema: {type: string}
        - name: X-JFrog-Event-Id
          in: header
          required: false
          schema: {type: string}
        - name: X-Correlation-Id
          in: header
          required: false
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JfrogWebhookPayload'
      responses:
        '202':
          description: The normalized event was durably accepted by EventBridge.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookReceipt'
        '400':
          description: Payload, repository, path, or event shape is invalid.
        '401':
          description: Signature, origin, or subscription is invalid.
        '413':
          description: Payload exceeds the configured maximum.
        '415':
          description: Content type is not application/json.
        '503':
          description: Signing is not configured or EventBridge did not accept the event.
...
components:
  securitySchemes:
    oidcSession:
      type: openIdConnect
      openIdConnectUrl: https://login.microsoftonline.com/common/v2.0/.well-known/openid-configuration
      description: |
        Microsoft Entra OIDC is terminated by the production ALB or by the local Spring OAuth2 client.
        The browser holds only a secure server-side session cookie; ALB identity headers are verified by the API.
  parameters:
...
    ApmId:
      name: apm_id
      in: query
      description: Optional selected APM context. Unknown or unauthorized values fail closed.
      schema: {type: string, pattern: '^APM[0-9]{7}$'}
...
  responses:
...
    BadRequest:
      description: A filter, cursor, route, or request value is invalid.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Unauthorized:
      description: No valid OIDC session or verified ALB identity is present.
...
    IdentityUnavailable:
      description: Microsoft Graph or another required identity component is unavailable; access fails closed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
          additionalProperties: false
      additionalProperties: false
...
    CatalogPage:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/CatalogPackage'}
        next_cursor: {type: [string, 'null']}
        total: {type: integer, format: int64, minimum: 0}
      additionalProperties: false
    CatalogCounts:
      type: object
      required: [modules, providers]
      properties:
        modules: {type: integer, format: int64, minimum: 0}
        providers: {type: integer, format: int64, minimum: 0}
      additionalProperties: false
    CatalogPackage:
      type: object
      required:
        - id
        - kind
        - namespace
        - name
        - target
        - title
        - description
        - latest_version
        - owners
        - support_level
        - lifecycle
        - verification
        - risk_tier
        - source_address
        - updated_at
        - versions
        - symbols
      properties:
        id: {type: string, description: Canonical module/... or provider/... identity.}
        kind: {type: string, enum: [module, provider]}
        namespace: {type: string}
        name: {type: string}
        target: {type: string, description: Module target; empty for providers.}
        title: {type: string}
        description: {type: string}
        latest_version:
          type: string
          description: Latest visible version, or the selected version on an explicit version route.
        owners: {type: array, items: {type: string}}
        support_level: {type: string, enum: [supported, maintenance, experimental, deprecated, revoked, archived]}
        lifecycle: {type: string, enum: [draft, candidate, approved, maintenance, deprecated, revoked, archived]}
        verification: {type: string}
        risk_tier: {type: string, enum: [low, medium, high, critical]}
        source_address: {type: string}
        updated_at: {type: string, format: date-time}
        versions:
          type: array
          items: {$ref: '#/components/schemas/PackageVersion'}
        symbols:
          type: array
          items: {$ref: '#/components/schemas/Symbol'}
      additionalProperties: false
    PackageVersion:
      type: object
      required:
        - version
        - published_at
        - package_digest
        - documentation_digest
        - documentation_root
        - source_commit
        - prerelease
        - deprecated
        - revoked
      properties:
        version: {type: string}
        published_at: {type: string, format: date-time}
        package_digest: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
        documentation_digest: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
        documentation_root: {type: string}
        source_commit: {type: string}
        prerelease: {type: boolean}
        deprecated: {type: boolean}
        revoked: {type: boolean}
      additionalProperties: false
    Symbol:
      type: object
      required: [kind, name, path]
      properties:
        kind: {type: string}
        name: {type: string}
        description: {type: [string, 'null']}
        path: {type: string}
      additionalProperties: false
...
    WebhookReceipt:
      type: object
      required: [event_id, publication_id]
      properties:
        event_id: {type: string}
        publication_id: {type: string}
      additionalProperties: false
    JfrogWebhookPayload:
      type: object
      description: Supported JFrog artifact or property webhook envelope. Unknown vendor fields are ignored.
      properties:
        id: {type: string}
        event_id: {type: string}
        event_type: {type: string}
        type: {type: string}
        occurred_at: {type: string, format: date-time}
        data:
          type: object
          properties:
            repo_key: {type: string}
            repository: {type: string}
            path: {type: string}
            item_path: {type: string}
            event_type: {type: string}
            properties:
              type: object
              additionalProperties: {type: string}
          additionalProperties: true
      additionalProperties: true
...